Ejercicio Servidor Web- Implementando Put

+++ EjercicioServidorWeb/src/app/componentes/tabla-usuarios/tabla-usuarios.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { UsuariosService } from 'src/app/services/usuarios.service';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-tabla-usuarios',
@@ -10,9 +11,10 @@ export class TablaUsuariosComponent implements OnInit {
 
   usuarios: any[] = [];
 
-  constructor( private service: UsuariosService) {
+  constructor( private service: UsuariosService, private router: Router) {
     this.service.getUsuario().subscribe( res => {
       this.usuarios = res['usuarios'];
+      console.log(this.usuarios);
     });
   }
 
@@ -30,9 +32,13 @@ export class TablaUsuariosComponent implements OnInit {
     const _id = this.usuarios[item]._id;
 
     console.log(_id);
-    this.service.deleteUsuario( _id ).subscribe( res => {
+    if (confirm('¿Estas seguro de eliminar este registro?')) {
+      console.log('Acepto');
+      this.service.deleteUsuario( _id ).subscribe( res => {
       console.log(res);
-    })
+      this.router.navigate(['tablaUsuarios'])
+      });
+    }   
   }
 
 }
